requirements.txt
google-cloud-bigquery
pandas
requests
functions-framework
db-dtypes